Revert "Merging to release-5.3: [TT-12355] added log format config, func and test cases (#6344)" (#6485)
### **User description**
Reverts TykTechnologies/tyk#6484


___

### **PR Type**
enhancement, bug fix


___

### **Description**
- Removed the `LogFormat` configuration option from the codebase and
related schema.
- Refactored and renamed initialization functions in the gateway server
code.
- Simplified log formatter setup by removing the JSONFormatter and
related functionality.
- Removed the `github.com/goccy/go-json` dependency from the project.


___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant
files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
<summary><strong>config.go</strong><dd><code>Remove LogFormat
configuration option</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

config/config.go

- Removed the `LogFormat` configuration option.



</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-fe44f09c4d5977b5f5eaea29170b6a0748819c9d02271746a20d81a5f3efca17">+0/-4</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
<summary><strong>server.go</strong><dd><code>Refactor and rename
initialization functions</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

gateway/server.go

<li>Renamed <code>initRPCCache</code> to
<code>InitializeRPCCache</code>.<br> <li> Renamed
<code>initSystem</code> to <code>initialiseSystem</code>.<br> <li>
Removed log format initialization.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-4652d1bf175a0be8f5e61ef7177c9666f23e077d8626b73ac9d13358fa8b525b">+21/-31</a>&nbsp;
</td>

</tr>                    

<tr>
  <td>
    <details>
<summary><strong>testutil.go</strong><dd><code>Update test setup
function name</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
</dd></summary>
<hr>

gateway/testutil.go

- Renamed `initSystem` to `initialiseSystem` in test setup.



</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-7aaf6ae49fb8f58a8c99d337fedd15b3e430dd928ed547e425ef429b10d28ce8">+1/-1</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
<summary><strong>log.go</strong><dd><code>Simplify log formatter setup
and remove JSONFormatter</code>&nbsp; &nbsp; &nbsp; &nbsp;
</dd></summary>
<hr>

log/log.go

<li>Removed JSONFormatter and related functionality.<br> <li> Simplified
log formatter setup.<br>


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-894630a7ed925c7768a861c4465bc9ad393f7937df7cac5c77ab123c03921aee">+43/-59</a>&nbsp;
</td>

</tr>                    
</table></td></tr><tr><td><strong>Configuration
changes</strong></td><td><table>
<tr>
  <td>
    <details>
<summary><strong>schema.json</strong><dd><code>Remove log_format option
from schema</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

cli/linter/schema.json

- Removed `log_format` configuration option from schema.



</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-103cec746d3e61d391c5a67c171963f66fea65d651d704d5540e60aa5d574f46">+0/-4</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>                    
</table></td></tr><tr><td><strong>Dependencies</strong></td><td><table>
<tr>
  <td>
    <details>
<summary><strong>go.mod</strong><dd><code>Remove go-json
dependency</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

go.mod

- Removed dependency on `github.com/goccy/go-json`.



</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6">+0/-1</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>                    

<tr>
  <td>
    <details>
<summary><strong>go.sum</strong><dd><code>Remove go-json
checksums</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

go.sum

- Removed checksums for `github.com/goccy/go-json`.



</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6485/files#diff-3295df7234525439d778f1b282d146a4f1ff6b415248aaac074e8042d9f42d63">+0/-2</a>&nbsp;
&nbsp; &nbsp; </td>

</tr>                    
</table></td></tr></tr></tbody></table>

___

> 💡 **PR-Agent usage**:
>Comment `/help` on the PR to get a list of all available PR-Agent tools
and their descriptions
