+++ cli/linter/schema.json
@@ -632,10 +632,6 @@
       "type": "string",
       "enum": ["", "debug", "info", "warn", "error"]
     },
-    "log_format": {
-      "type": "string",
-      "enum": ["", "standard", "json"]
-    },
     "enable_http_profiler": {
       "type": "boolean"
     },
